backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.1
paho-mqtt==1.6.1
paramiko==3.3.1
pathspec==0.11.2
platformdirs==3.10.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tencentcloud-sdk-python==3.0.989
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.5
yarl==1.9.2
